// src/module/courses/dto/create-course-section-lesson.dto.ts
import { CreateSectionDto } from "src/module/sections/dto/create-section.dto";

export class CreateCourseSectionLessonDto {
  courseId: string;
  title: string;
  description: string;
  ratingScore: number;
  ratingNumber: number;
  image: string;
  price: number;
  sale: number;
  authorId: string;
  category: string[];
  sections: CreateSectionDto[];
}